consolidate stripe connect logic into single hook
