+++ spa/src/components/dashboard/Dashboard.js
@@ -1,6 +1,4 @@
-import { useState, useEffect } from 'react';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { Route, Switch, Redirect, useLocation, useHistory } from 'react-router-dom';
+import { Route, Switch, Redirect, useLocation } from 'react-router-dom';
 
 import * as S from './Dashboard.styled';
 
@@ -25,22 +23,14 @@ import { CONTENT_SECTION_ACCESS_FLAG_NAME } from 'constants/featureFlagConstants
 import flagIsActiveForUser from 'utilities/flagIsActiveForUser';
 import hasContributionsDashboardAccessToUser from 'utilities/hasContributionsDashboardAccessToUser';
 import { usePageContext } from './PageContext';
-import { USER_ROLE_ORG_ADMIN_TYPE } from 'constants/authConstants';
-import axios from 'ajax/axios';
-import { getStripeAccountLinkCreateCompletePath } from 'ajax/endpoints';
 import useUser from 'hooks/useUser';
-
-const postAccountLinkSuccess = (rpId) => {
-  return axios.post(getStripeAccountLinkCreateCompletePath(rpId));
-};
+import useConnectStripeAccount from 'hooks/useConnectStripeAccount';
 
 function Dashboard() {
   const { flags } = useFeatureFlags();
   const { page } = usePageContext();
   const { user } = useUser();
-
-  const history = useHistory();
-  const [revenueProgramIdForVerification, setRevenueProgramIdForVerification] = useState();
+  const { requiresStripeVerification } = useConnectStripeAccount();
 
   const hasContributionsSectionAccess = user?.role_type && hasContributionsDashboardAccessToUser(flags);
   const hasContentSectionAccess = user?.role_type && flagIsActiveForUser(CONTENT_SECTION_ACCESS_FLAG_NAME, flags);
@@ -51,75 +41,12 @@ function Dashboard() {
     ? DONATIONS_SLUG
     : 'not-found';
 
-  const { pathname, search } = useLocation();
+  const { pathname } = useLocation();
   const isEditPage = pathname.includes(EDITOR_ROUTE);
-
-  /*If user is an rp_admin whose first RP is not verified, we set the id
-  of that RP to trigger display of Stripe Connect CTA elements.
-
-  Note that the logic here is unsophisticated. It presumes that the primary
-  user (for now) is an rp_admin with a single revenue program. It will not
-  cause the Stripe Connect CTA elements to be displayed to users with other role types
-  than rp_admin. It also does not support connecting > 1 revenue program beyond
-  the first to appear in user.revenue_programs.
-  */
-  useEffect(() => {
-    if (
-      user?.role_type?.[0] === USER_ROLE_ORG_ADMIN_TYPE &&
-      user?.revenue_programs?.[0]?.payment_provider_stripe_verified === false
-    ) {
-      setRevenueProgramIdForVerification(user.revenue_programs[0].id);
-    }
-  }, [user?.revenue_programs, user?.role_type]);
-
-  /*After a user goes through the Stripe Account Link on the Stripe site,
-  they will be sent back to the dashboard root but with an empty stripeAccountLinkSuccess
-  query param, which signals that we should tell the server to retrieve the Stripe Account
-  and update the `stripe_verified` property on the RP's payment provider if the Stripe Account
-  is properly configured.
-  */
-  const queryClient = useQueryClient();
-  const signalStripeAccountLinkSuccess = useMutation((rpId) => postAccountLinkSuccess(rpId), {
-    onSuccess: () => {
-      // NB: `['user']` is meant to refer to a named query generated by useQuery call
-      // in custom useUser hook. This will cause the user to be refetched,
-      // the user's revenue program should now appear as having Stripe verified, which will
-      // in turn hide the Stripe Account Link CTAs.
-      queryClient.invalidateQueries(['user']);
-      setRevenueProgramIdForVerification(null);
-      // this removes the `stripeAccountLinkSuccess` query param from route that
-      // tells app to ping stripe account link success endpoint on server.
-      history.replace(pathname);
-    }
-  });
-
-  /* After a user goes through the Stripe Account onboarding flow on the Stripe site,
-  they get sent back to /stripe-account-link-complete, which (in DashboardRouter.js) redirects
-  to the dashboard app route with a `stripeAccountLinkSuccess` present. If this component
-  loads and that param is present, we signal to the server that Stripe Account Link is complete,
-  and the server will update the revenue program's payment provider's stripe_verified value
-  to be `true` if it sees that the account is ready to accept charges.
-  */
-  useEffect(() => {
-    const params = new URLSearchParams(search);
-    if (
-      revenueProgramIdForVerification &&
-      [...params.keys()].includes('stripeAccountLinkSuccess') &&
-      !signalStripeAccountLinkSuccess.isLoading &&
-      !signalStripeAccountLinkSuccess.isSuccess &&
-      !signalStripeAccountLinkSuccess.isError
-    ) {
-      signalStripeAccountLinkSuccess.mutate(revenueProgramIdForVerification);
-    }
-  }, [revenueProgramIdForVerification, search, signalStripeAccountLinkSuccess]);
 
   return (
     <S.Outer>
-      {revenueProgramIdForVerification ? (
-        <ConnectStripeElements revenueProgramId={revenueProgramIdForVerification} />
-      ) : (
-        ''
-      )}
+      {requiresStripeVerification ? <ConnectStripeElements /> : ''}
       <DashboardTopbar isEditPage={isEditPage} page={page} />
       <S.Dashboard data-testid="dashboard">
         {isEditPage ? null : <DashboardSidebar />}
